backend/requirements.txt
alembic==1.14.1
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.0.1
blinker==1.8.2
certifi==2024.12.14
click==8.1.8
deprecation==2.1.0
exceptiongroup==1.2.2
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
gotrue==1.3.1
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==2.1.5
packaging==24.2
postgrest==0.13.2
psycopg2-binary==2.9.9
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
realtime==1.0.6
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
storage3==0.6.1
StrEnum==0.4.15
supabase==2.0.3
supafunc==0.3.3
typing_extensions==4.12.2
websockets==12.0
Werkzeug==2.3.7
zipp==3.20.2
